add container voulme mount setting to test
